lost-found-reunion/requirements.txt
streamlit
sentence-transformers
transformers
chromadb
openai
pandas
Pillow
requests
beautifulsoup4
scikit-learn
torch
torchvision
numpy
tqdm
lxml
